requirements.txt
huggingface-hub
numpy
onnxruntime
pandas
pillow
tqdm
requests